using streams instead of synchronus file reading
For performance improvement.
Because streams read file in chunks while readFileSync is a blocking
function and that's why it will wait until the whole file is read and
loaded it to the memory after then.

This approach will significantly improve performanace as the log file is
huge (more than 3.5 MBs)